http: extract transport/error layer from source
Move SourceError, _GITHUB_HOSTS, _auth_headers, _rate_limit_hint,
_http_error, _GH_JSON, and _get from source.py into a new leaf module
http.py. Re-export all seven names from source.py so existing callers
resolve source.SourceError, source._get, etc. without change.

Add a timeout keyword parameter (default 30 s) to _get so the new
archive backend can pass a longer timeout without touching existing
callers.

Full test suite (304 tests) passes unchanged.
